update to fix errors
fixed a whole bunch of errors regarding the code searching for values that don't exist in case you didnt enter enough characters in the search bar, say for example E280 and no mdid, tmn, xtid header. in this case it's searching for values in memory spaces that don't exist, so it crashes. should not happen anymore

fixed a decoding error regarding the lock bit segment, mixed up MSB and LSB, and the check to enable decoding was done on the wrong value ( lock_bit_segment_present when it should have been S==1 ).